requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
cvlib==0.2.7
decorator==4.4.2
easyocr==1.7.2
filelock==3.18.0
fsspec==2025.5.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imutils==0.5.4
Jinja2==3.1.6
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
proglog==0.1.12
progressbar==2.5
pyclipper==1.3.0.post6
pydub==0.25.1
python-bidi==0.6.6
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
sympy==1.14.0
tifffile==2025.5.10
tiktoken==0.9.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
